requirements.txt
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1751974613617/work
beautifulsoup4==4.13.5
chardet==5.2.0
cssselect==1.3.0
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1756937038999/work
django-cors-headers==4.9.0
djangorestframework==3.16.1
lxml==6.0.1
lxml_html_clean==0.4.2
markdownify==1.2.0
readability-lxml==0.8.4.1
six==1.17.0
soupsieve==2.8
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1734007102193/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
